+++ toolium/test/pageelements/test_page_elements_groups.py
@@ -108,6 +108,15 @@ def test_reset_object_page_elements_groups(driver_wrapper):
     assert column_21.input._web_element is not None
     assert column_21.link._web_element is not None
     assert column_21.input_with_parent._web_element is not None
+    # Check that the group elements have the group as parent
+    assert column_11.parent == row_1
+    assert column_21.parent == row_2
+    assert column_11.input.parent == column_11
+    assert column_11.link.parent == column_11
+    assert column_11.input_with_parent.parent == column_11
+    assert column_21.input.parent == column_21
+    assert column_21.link.parent == column_21
+    assert column_21.input_with_parent.parent == column_21
 
     table_page.reset_object()
 
@@ -128,3 +137,12 @@ def test_reset_object_page_elements_groups(driver_wrapper):
     assert column_21.input._web_element is None
     assert column_21.link._web_element is None
     assert column_21.input_with_parent._web_element is None
+    # Check that the group elements have the group as parent
+    assert column_11.parent == row_1
+    assert column_21.parent == row_2
+    assert column_11.input.parent == column_11
+    assert column_11.link.parent == column_11
+    assert column_11.input_with_parent.parent == column_11
+    assert column_21.input.parent == column_21
+    assert column_21.link.parent == column_21
+    assert column_21.input_with_parent.parent == column_21